news_eval_backend/requirements.txt
Django>=5.2
djangorestframework>=3.16
django-cors-headers>=4.9

gunicorn>=22.0
psycopg[binary]>=3.2

dj-database-url>=2.2
whitenoise>=6.7
